+++ packages/codebytes/src/theme.d.ts
@@ -1,0 +1,4 @@
+import { CoreTheme } from '@codecademy/gamut-styles';
+declare module '@emotion/react' {
+  export interface Theme extends CoreTheme {}
+}